Move all response logic into PeerOperations
